refactor: move tests from __tests__ to src directory, fix AWS SDK Filter properties
